requirements.txt
langchain~=0.2
langgraph~=0.0.30
langserve~=0.0.20
weaviate-client~=4.5
neo4j~=5.17
graphiti~=0.4
fastapi~=0.111
uvicorn[standard]~=0.29
python-dotenv~=1.0
pydantic~=2.7
